# backend/src/main/resources/openapi/api.yaml
openapi: 3.0.1
info:
  title: Homni Togli API
  version: '2.0.0'
  description: Per-project feature toggle management with RBAC and API key authentication.
servers:
  - url: 'http://localhost:8080'
    description: local
security:
  - apiKey: []
  - oidc: []
...
paths:
...
  /projects:
    post:
      tags: [Projects]
      operationId: createProject
      summary: Create a new project
      description: Requires PLATFORM_ADMIN role.
      security:
        - oidc: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSingleResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '409':
          $ref: '#/components/responses/409'
        '500':
          $ref: '#/components/responses/500'

    get:
      tags: [Projects]
      operationId: listProjects
      summary: List projects
      description: |
        PLATFORM_ADMIN sees all projects (including archived). Regular users see projects
        they are a member of, but archived projects only when they have the ADMIN role on them.
        Each item carries pre-aggregated `togglesCount`, `environmentsCount` and `membersCount`
        for the project list UI; the `pagination` envelope reflects the filtered query, while
        `totalCount` / `archivedCount` reflect the workspace-wide visible totals (independent
        of the search/archived query parameters) so the page header can show "N projects · M
        archived" without an extra round-trip.
      security:
        - oidc: []
      parameters:
        - name: q
          in: query
          required: false
          description: Case-insensitive substring match against project name and slug.
          schema:
            type: string
        - name: archived
          in: query
          required: false
          description: |
            Tri-state filter applied on top of the visibility rules above. Omit
            (or pass `null`) to return both active and archived projects, `true`
            to return only archived ones, `false` to return only active ones.
          schema:
            type: boolean
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'
  /projects/by-slug/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
        description: Unique project slug

    get:
      tags: [Projects]
      operationId: getProjectBySlug
      summary: Get a project by slug
      description: |
        Returns the project matching the slug. Returns 404 if the slug does
        not exist and 403 if the caller is not a member of the project.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSingleResponse'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/projectId'

    patch:
      tags: [Projects]
      operationId: updateProject
      summary: Update project fields (name, description, archived status)
      description: |
        Requires MANAGE_MEMBERS permission. Set archived=true to freeze the project
        (which also bulk-disables every enabled feature toggle inside it). Once a project
        is archived, the only allowed change is unarchive (archived=false): any attempt
        to update name or description on an archived project is rejected with 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSingleResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '409':
          $ref: '#/components/responses/409'
        '500':
          $ref: '#/components/responses/500'

  # ========== Feature Toggles ==========
...
  /users/me:
    get:
      tags: [Users]
      operationId: getCurrentUser
      summary: Get current authenticated user
      description: Returns the profile of the currently authenticated OIDC user.
      security:
        - oidc: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSingleResponse'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List all platform users
      description: Requires PLATFORM_ADMIN role.
      security:
        - oidc: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
  /users/search:
    get:
      tags: [Users]
      operationId: searchUsers
      summary: Search users by email or name
      description: Returns up to 20 users matching the query. Available to any authenticated user (for member assignment).
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          description: Search query (matches email or name, case-insensitive)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    patch:
      tags: [Users]
      operationId: updateUser
      summary: Update user platform role or status
      description: Requires PLATFORM_ADMIN role. Use to promote/demote or disable/activate users.
      security:
        - oidc: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSingleResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '409':
          $ref: '#/components/responses/409'
        '500':
          $ref: '#/components/responses/500'
components:
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '404':
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '409':
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ResponseMeta:
      type: object
      required: [timestamp]
      properties:
        timestamp:
          type: string
          format: date-time
    Pagination:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      required: [payload, meta]
      properties:
        payload:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'

    # ---------- Project ----------
    CreateProjectRequest:
      type: object
      required: [slug, name]
      properties:
        slug:
          type: string
          minLength: 2
          maxLength: 50
          pattern: '^[A-Za-z][A-Za-z0-9_-]*$'
          example: MY_PROJECT
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: My Project
        description:
          type: string
          maxLength: 1000
        environments:
          type: array
          description: |
            Subset of the platform-configured default environments to bootstrap
            in the new project. Omit the field entirely to create the project
            with all configured defaults; pass an empty array to opt out
            and create the project without any environments. Names that are
            not in the configured defaults are rejected with 422.
          items:
            type: string
            maxLength: 50
          example: [DEV, PROD]
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        archived:
          type: boolean
          description: Set to true to archive (readonly), false to unarchive
    Project:
      type: object
      required: [id, slug, name, archived, createdAt, togglesCount, environmentsCount, membersCount]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        archived:
          type: boolean
        createdAt:
          type: string
          format: date-time
        myRole:
          type: string
          enum: [ADMIN, EDITOR, READER]
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        togglesCount:
          type: integer
          format: int64
          description: Number of feature toggles in this project.
        environmentsCount:
          type: integer
          format: int64
          description: Number of environments in this project.
        membersCount:
          type: integer
          format: int64
          description: Number of members in this project.
    ProjectSingleResponse:
      type: object
      required: [payload, meta]
      properties:
        payload:
          $ref: '#/components/schemas/Project'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ProjectListResponse:
      type: object
      required: [payload, pagination, totalCount, archivedCount, meta]
      properties:
        payload:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        pagination:
          $ref: '#/components/schemas/Pagination'
        totalCount:
          type: integer
          format: int64
          description: |
            Total number of projects visible to the caller in the workspace,
            ignoring the `q` and `archived` query parameters. Used by the
            page header subtitle ("N projects · M archived").
        archivedCount:
          type: integer
          format: int64
          description: |
            Number of archived projects visible to the caller in the workspace,
            ignoring the `q` and `archived` query parameters.
        meta:
          $ref: '#/components/schemas/ResponseMeta'

    # ---------- Feature Toggle ----------
...
    User:
      type: object
      required: [id, email, platformRole, active, createdAt]
      properties:
        id:
          type: string
          format: uuid
        oidcSubject:
          type: string
        email:
          type: string
        name:
          type: string
        platformRole:
          type: string
          enum: [PLATFORM_ADMIN, USER]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
    UpdateUserRequest:
      type: object
      properties:
        platformRole:
          type: string
          enum: [PLATFORM_ADMIN, USER]
        active:
          type: boolean
    UserSingleResponse:
      type: object
      required: [payload, meta]
      properties:
        payload:
          $ref: '#/components/schemas/User'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    UserListResponse:
      type: object
      required: [payload, pagination, meta]
      properties:
        payload:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
    oidc:
      type: http
      scheme: bearer
      bearerFormat: JWT